refactor(editor): split editor sub-components & modals under max-lines-per-function (#196)
- EditorControls.tsx: extract reusable TrimRangeSlider (used for start/end)
- EditorTopbar.tsx: extract useProfileImage hook + EditorTopbarUserMenu
- ZoomEffectsPopup.tsx: extract useZoomEffectForm hook + ZoomEffectForm
  and ZoomEffectList sub-components
- ExportSettingsModal.tsx: move EXEMPT_EMAILS + estimateFileSize to module
  scope, extract ExportQualityFrameRate + ExportCompression rows
- CustomVideoControls.tsx: extract VideoUtilityControls cluster
